requirements.txt
Django==1.11
django-appconf==1.0.2
django-colorfield==0.1.14
django-grappelli==2.10.1
django-imagekit==4.0.2
django-jet==1.0.6
django-model-utils==3.1.1
djangorestframework==3.7.3
google-api-python-client==1.4.1
httplib2==0.10.3
oauth2client==4.1.2
olefile==0.44
pilkit==2.0
Pillow==4.3.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
rsa==3.4.2
singlemodeladmin==0.8
six==1.11.0
uritemplate==3.0.0
